Moved FormatLand out of every socket new land In order to do the rendering async with map Added doFilter which nan says does nothing
